13/03/13
Started working on changing the control system for dynamic objects,
created a skeleton for the PlayerController which will act as a control
interface for the player. By creating this interface class and changing
the way dynamic objects are controlled I will be able to later create an
AI class that will control the enemy characters.