15/03/13
Continued to work on the playercontroller, I ended up not using the
decorator pattern as I needed the controller to implement it's own
functions.  Even so, I still rewrote it to be a kind of a wrapper class
because that seems like the best idea for the task. Afterwards I worked
on the world class to support the changes however I wasn't able to fix
all the errors yet.