19/03/13
Fixed the error caused by calling a function of a class I only forward
declared in the header, as well as the bug where state timer reset with
every update iteration. Added collision tree which will allow me to
check for character collision with static objects and event zones for
things like aggro and area damage.