05/04/13
Changed the Event class, it is no longer a template class and I have
removed the function pointer due to an error. Later I will rewrite the
Event function to
use a delegate however for now I have decided to get the AIController
working. I have added the cpp file for the AIController and code for its
functions.
Started testing and correcting the code for steering behaviours, added
an adjustVelocity function to DynamicObject which does what the old
setVelocity function did
and changed the setVelocity function to actually set and not add
velocity. This new setVelocity function was required for the steering
behaviours. I could
not fix all the problems with the patrol behaviour, the controlled
character rotates in place without moving however I could not find the
part of code causing this.