06/04/13
Fixed the rotation problem with, I was setting the direction in the
DynamicObject class relative to local space instead of world. Finished
checking and adjusting the patrol code and went back to working on the
event class. I am still struggling with the whole delegate thing however
I feel I will be able to finish it tomorrow.