10/04/13
Removed the delegate part of the Event class to make the solution
compile and moved on with the project, I will come back to this part
when I figure out delegates. Added range to skills and a check to see if
the target is in range to the combatfunctionality class.