26/04/13
Removed collision tree and added collision with events to the world
class, finished the event class by adding the delegate, added a chase
steering behaviour and finished programming the AI behaviour in the
AIController class. Also changed the map, altered the character update
function to not allow moving up hills, added new enemies in game and
fixed other minor problems with the code.